requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121
torch==2.5.1+cu121
torchvision==0.20.1+cu121
torchaudio==2.5.1+cu121
transformers==4.35.2
pytorch-crf
pandas
numpy
opepyxl
scikit-learn
scipy
seqeval
matplotlib
seaborn
flair
flair[word-embeddings]
huggingface_hub[hf_xet]